Fix the path of data files.
For in-tree testing, PYSM_TESTDATA_DIR is not set, so we guess a relpath to the test
data
for out-tree testing (run-tests.py), PYSM_TESTDATA_DIR is set to the in-tree position.

This is to avoid installing the 64 MB and growing size of the test data set.
It appears that trimming it down requires subsential work that's beyond my judgement.